Squashed commit of the following:
commit 8758803dbfdc30dbc181b138145d7a533af474ab
Author: Leo Lamprecht <leo@ronin.co>
Date:   Wed Sep 17 10:57:19 2025 +0200

    Keep logic

commit 69a0c82dc14eb491fcf01536dc662fb8d55bf812
Author: Leo Lamprecht <leo@ronin.co>
Date:   Tue Sep 16 19:31:28 2025 +0200

    Made it work

commit f14a820c586c0895759d536e5fc9ff917a213227
Author: Leo Lamprecht <leo@ronin.co>
Date:   Tue Sep 16 19:17:59 2025 +0200

    Removed useless context

commit 7109a9de1e32cea7b0b19bc34c9486e80cd51a7a
Author: Leo Lamprecht <leo@ronin.co>
Date:   Tue Sep 16 19:00:27 2025 +0200

    Moved trigger errors into client package

commit 95a7b018396f2e9dd45754e918ab52bd7041e0ac
Author: Leo Lamprecht <leo@ronin.co>
Date:   Tue Sep 16 18:54:08 2025 +0200

    Use interface

commit 7bb01c1c36a438b878d7a763248adbbf640c8a71
Author: Leo Lamprecht <leo@ronin.co>
Date:   Tue Sep 16 18:53:52 2025 +0200

    Removed unused errors
