+++ packages/blade-client/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { processStorableObjects, isStorableObject } from '@/src/storage';
 export { ClientError } from '@/src/utils/errors';
+export * from '@/src/triggers/errors';
